Finished the AJAX live updates
- Used some nasty query comparisons to find differences in the arrays and update things on the page
- Added a function which dumps all current known CIDR objects in a JSON form to the browser
- Change an error message

Additional notes:
I am sure this code is awful for the time being but I hope to clean up a bit in the next few commits.
I am not sure how to get jquery calls that directly affect the page to be put into another JS page to
clean things up a bit in the index.html template but I might find a solution.
Some of the functions and variables are not normalized (I am very tired and not feeling well so I kinda
made a mess of everything) so I might have spacing issues and general formatting might look a bit weird.